add some commonly used protein prep items to conftest instead
