+++ asapdiscovery-modeling/asapdiscovery/modeling/tests/test_schema.py
@@ -6,45 +6,32 @@ from asapdiscovery.data.openeye import oechem
 from asapdiscovery.data.schema import CrystalCompoundData
 from asapdiscovery.data.testing.test_resources import fetch_test_file
 from asapdiscovery.modeling.modeling import save_design_unit
-from asapdiscovery.modeling.schema import MoleculeFilter, PreppedTarget
+from asapdiscovery.modeling.schema import MoleculeFilter, PreppedTarget, PreppedTargets
 
 
-@pytest.mark.parametrize(
-    "components",
-    [["ligand"], ["protein", "ligand"], ["protein", "ligand", "water"]],
-)
-def test_molecule_filter_component_success(components):
-    MoleculeFilter(components_to_keep=components)
-
-
-@pytest.mark.parametrize(
-    "components",
-    [["prot"], ["basket", "ligand"], "help"],
-)
-def test_molecule_filter_component_failure(components):
-    with pytest.raises(ValueError):
+class TestMoleculeFilter:
+    @pytest.mark.parametrize(
+        "components",
+        [["ligand"], ["protein", "ligand"], ["protein", "ligand", "water"]],
+    )
+    def test_molecule_filter_component_success(self, components):
         MoleculeFilter(components_to_keep=components)
 
+    @pytest.mark.parametrize(
+        "components",
+        [["prot"], ["basket", "ligand"], "help"],
+    )
+    def test_molecule_filter_component_failure(self, components):
+        with pytest.raises(ValueError):
+            MoleculeFilter(components_to_keep=components)
 
-@pytest.mark.parametrize(
-    "ligand_chain",
-    [list("A")],
-)
-def test_molecule_filter_ligand_chain_failure(ligand_chain):
-    with pytest.raises((ValueError, pydantic.ValidationError)):
-        MoleculeFilter(ligand_chain=ligand_chain)
-
-
-# This needs to have a scope of session so that a new tmp file is not created for each test
-@pytest.fixture(scope="session")
-def output_dir(tmp_path_factory, local_path):
-    if not type(local_path) == str:
-        return tmp_path_factory.mktemp("test_prep")
-    else:
-        local_path = Path(local_path)
-        local_path.mkdir(exist_ok=True)
-        assert local_path.exists()
-        return local_path
+    @pytest.mark.parametrize(
+        "ligand_chain",
+        [list("A")],
+    )
+    def test_molecule_filter_ligand_chain_failure(self, ligand_chain):
+        with pytest.raises((ValueError, pydantic.ValidationError)):
+            MoleculeFilter(ligand_chain=ligand_chain)
 
 
 def test_prepped_target(output_dir):
@@ -67,3 +54,54 @@ def test_prepped_target(output_dir):
     for fn in [saved_target.ligand, saved_target.complex, saved_target.protein]:
         assert Path(fn).exists()
         assert Path(fn).is_file()
+
+
+class TestPreppedTargets:
+    def test_dataset_creation(self, target_dataset, sars_target, mers_target):
+        assert len(target_dataset.iterable) == 2
+        assert target_dataset.iterable[0].source.str_fn == str(
+            sars_target.source.str_fn
+        )
+        assert target_dataset.iterable[1].source.str_fn == str(
+            mers_target.source.str_fn
+        )
+
+    @pytest.mark.skip(
+        reason="Multiple embedded schema objects need more logic to serialize to csv"
+    )
+    def test_dataset_csv_usage(
+        self, target_dataset, output_dir, csv_name="to_prep.csv"
+    ):
+        to_prep_csv = output_dir / csv_name
+        target_dataset.to_csv(to_prep_csv)
+        assert to_prep_csv.exists()
+        assert to_prep_csv.is_file()
+
+        dataset = PreppedTargets.from_csv(output_dir / csv_name)
+        assert dataset == target_dataset
+
+        dataset.iterable[0].active_site_chain = "B"
+        assert dataset != target_dataset
+
+    def test_dataset_pickle(self, target_dataset, output_dir, pkl_name="to_prep.pkl"):
+        pkl_file = output_dir / pkl_name
+        target_dataset.to_pkl(pkl_file)
+        assert pkl_file.exists()
+        assert pkl_file.is_file()
+
+        loaded_dataset = PreppedTargets.from_pkl(pkl_file)
+
+        assert loaded_dataset == target_dataset
+
+    @pytest.mark.skip(
+        reason="Multiple embedded schema objects need more logic to serialize to json"
+    )
+    def test_dataset_json(self, target_dataset, output_dir, json_name="to_prep.json"):
+        json_file = output_dir / json_name
+        target_dataset.to_json(json_file)
+        assert json_file.exists()
+        assert json_file.is_file()
+
+        loaded_dataset = PreppedTargets.from_json(json_file)
+
+        assert loaded_dataset == target_dataset
